adjusting details for elections, cms edit
