+++ src/content/navbar.json
@@ -66,6 +66,10 @@
         {
           "text": "Join OpenInfra Foundation ",
           "url": "/join/"
+        },
+        {
+          "text": "Members",
+          "url": "/a/community/members"
         }
       ]
     },
